+++ src/audiotree/__init__.py
@@ -3,7 +3,6 @@ __author__ = "David Braun"
 # Effort-based versioning. Don't move the line above. It must be the first line due to `docs/source/conf.py`
 from .core import AudioTree
 from .core import SaliencyParams
-from .core import batch_audiotrees
 from .writer import AudioWriter
 from .memmap_writer import MemmapWriter, FieldSpec
 from . import sources
@@ -12,7 +11,6 @@ from . import transforms
 __all__ = [
     "AudioTree",
     "SaliencyParams",
-    "batch_audiotrees",
     "AudioWriter",
     "MemmapWriter",
     "FieldSpec",