+++ src/audiotree/transforms/core.py
@@ -8,6 +8,7 @@ All other transforms are in the function-based API in functional.py.
 
 from typing import Any, Sequence
 
+from jax import numpy as jnp
 import numpy as np
 
 from grain._src.core import tree_lib
@@ -46,15 +47,25 @@ class Batch(BatchOperation):
 
     def _default_batch_fn(self, input_records: Sequence[Any]):
         """Batches records together and copies Numpy arrays to Shared Memory."""
+        from audiotree.core import AudioTree, _batch_audiotrees
+
         self._validate_structure(input_records)
 
         def stacking_function(*args):
-            first_arg = np.asanyarray(args[0])
+            first_arg = args[0]
+            if isinstance(first_arg, AudioTree):
+                return _batch_audiotrees(args)
+            if not isinstance(first_arg, (np.ndarray, jnp.ndarray)):
+                return list(args)
+            first_arg = np.asanyarray(first_arg)
             shape, dtype = (len(args) * first_arg.shape[0],) + first_arg.shape[1:], first_arg.dtype
             if not self._use_shared_memory or dtype.hasobject:
                 return np.concatenate(args, axis=0)
             return np.concatenate(args, axis=0, out=SharedMemoryArray(shape, dtype=dtype)).metadata
 
         return tree_lib.map_structure(
-            stacking_function, input_records[0], *input_records[1:]
+            stacking_function,
+            input_records[0],
+            *input_records[1:],
+            is_leaf=lambda x: isinstance(x, AudioTree),
         )